Fixed more annoyances

+++ src/PandarGeneralRaw/include/pandarGeneral/pandarGeneral_internal.h
@@ -363,7 +363,7 @@ typedef struct PacketsBuffer_s {
     m_iterPush++;
 
     if (m_iterPush == m_iterCalc) {
-      printf("buffer don't have space!,%d\n", m_iterPush - m_buffers.begin());
+      printf("buffer don't have space!,%d\n", (int)(m_iterPush - m_buffers.begin()));
       return 0;
     }
     if(((m_iterPush - m_iterCalc) > MAX_ITERATOR_DIFF) ||
